[RPD-195] Add _validation to core (#144)
* Move functions to core _validation

* Move matcha state path to constants

* Move functions to core _validation

* Move matcha state path to constants

* Move region validation to core and update mocks

* Remove unnecessary constants
