+++ src/matcha_ml/cli/_validation.py
@@ -1,83 +1,12 @@
-"""Validation for user inputs."""
-import json
-import os
+"""Validation for cli inputs."""
 from difflib import get_close_matches
 from typing import List, Optional, Set, Union
 
-from azure.mgmt.confluent.models._confluent_management_client_enums import (  # type: ignore [import]
-    ProvisionState,
-)
 from typer import BadParameter
 
-from matcha_ml.cli.ui.print_messages import print_error
+from matcha_ml.core._validation import _is_valid_prefix, is_valid_region
 from matcha_ml.errors import MatchaInputError
 from matcha_ml.services import AzureClient
-
-# TODO: dynamically set both of these variables
-LONGEST_RESOURCE_NAME = "artifactstore"
-MAXIMUM_RESOURCE_NAME_LEN = 24
-MATCHA_STATE_PATH = os.path.join(".matcha", "infrastructure", "matcha.state")
-
-
-def _is_alphanumeric(prefix: str) -> bool:
-    """Check whether the prefix is an alphanumeric string.
-
-    Args:
-        prefix (str): the prefix to be checked.
-
-    Returns:
-        bool: True if it is an alphanumeric string; False if not.
-    """
-    return prefix.isalnum()
-
-
-def _check_length(prefix: str) -> bool:
-    """Check whether the prefix is the correct length.
-
-    Args:
-        prefix (str): the prefix to be checked.
-
-    Returns:
-        bool: True if the prefix is less than the maximum length; False if not.
-    """
-    return (len(prefix) + len(LONGEST_RESOURCE_NAME)) < MAXIMUM_RESOURCE_NAME_LEN
-
-
-def _is_not_digits(prefix: str) -> bool:
-    """Check whether the prefix is only numbers.
-
-    Args:
-        prefix (str): the prefix to be checked.
-
-    Returns:
-        bool: True if the prefix doesn't contain only numbers; False otherwise.
-    """
-    return not prefix.isdigit()
-
-
-PREFIX_RULES = {
-    "numbers": {
-        "func": _is_not_digits,
-        "message": "Resource group name prefix cannot contain only numbers.",
-    },
-    "alphanumeric": {
-        "func": _is_alphanumeric,
-        "message": "Resource group name prefix can only contain alphanumeric characters.",
-    },
-    "length": {
-        "func": _check_length,
-        "message": f"Resource group name prefix must be between 3 and {MAXIMUM_RESOURCE_NAME_LEN - len(LONGEST_RESOURCE_NAME)} characters long.",
-    },
-}
-
-
-def get_azure_client() -> AzureClient:
-    """Fetch an Azure Client.
-
-    Returns:
-        AzureClient: the azure client interface.
-    """
-    return AzureClient()
 
 
 def find_closest_matches(
@@ -111,11 +40,10 @@ def region_validation(region: str) -> str:
     Returns:
         str: the region if valid
     """
-    azure_client = get_azure_client()
-
-    if azure_client.is_valid_region(region):
+    if is_valid_region(region):
         return region
     else:
+        azure_client = AzureClient()
         closest = find_closest_matches(region, azure_client.fetch_regions())
 
         if closest:
@@ -147,25 +75,6 @@ def region_typer_callback(region: str) -> str:
         raise BadParameter(str(e))
 
     return region
-
-
-def _is_valid_prefix(prefix: str) -> str:
-    """Check for whether a prefix is valid.
-
-    Args:
-        prefix (str): the prefix to check.
-
-    Raises:
-        MatchaInputError: raised when the prefix is invalid.
-
-    Returns:
-        str: if valid, the prefix is returned.
-    """
-    for rule, checker in PREFIX_RULES.items():
-        if not checker["func"](prefix):  # type: ignore
-            raise MatchaInputError(checker["message"])
-
-    return prefix
 
 
 def prefix_typer_callback(prefix: str) -> str:
@@ -202,7 +111,7 @@ def prefix_typer_callback(prefix: str) -> str:
     except MatchaInputError as e:
         raise BadParameter(str(e))
 
-    azure_client = get_azure_client()
+    azure_client = AzureClient()
 
     if not azure_client.is_valid_resource_group(prefix):
         raise BadParameter(
@@ -210,40 +119,6 @@ def prefix_typer_callback(prefix: str) -> str:
         )
 
     return prefix
-
-
-def check_current_deployment_exists() -> bool:
-    """Checks the current deployment using the .matcha directory current contents if it exists.
-
-    Specifically, it checks whether the resource group exists on Azure.
-
-    Returns:
-        bool: True if a deployment currently exists, else False.
-    """
-    if not os.path.isfile(MATCHA_STATE_PATH):
-        return False
-
-    with open(MATCHA_STATE_PATH) as f:
-        data = json.load(f)
-
-    # Check if a resource group name prefix is present in matcha.state file
-    if data.get("cloud") is not None and "prefix" in data.get("cloud"):
-        resource_group_name = data["cloud"]["prefix"] + "-resources"
-
-        client = get_azure_client()
-        rg_state = client.resource_group_state(resource_group_name)
-
-        if rg_state is None:
-            return False
-        elif rg_state == ProvisionState.SUCCEEDED:
-            return True
-        else:
-            print_error(
-                f"Error, resource group '{resource_group_name}' is currently in the state '{rg_state.value}' which is currently not handled by matcha. Please check your resources on Azure."
-            )
-            return True
-    else:
-        return False
 
 
 def get_command_validation(argument: str, valid_options: List[str], noun: str) -> None:
