+++ tests/test_state/test_remote_state_manager.py
@@ -73,7 +73,6 @@ def broken_config_testing_directory(matcha_testing_directory: str) -> str:
         str: temporary working directory path that the configuration was written to
     """
     config_path = os.path.join(matcha_testing_directory, DEFAULT_CONFIG_NAME)
-    print(config_path)
     content = {}
     with open(config_path, "w") as f:
         json.dump(content, f)
